move optimize_bids into its own service
