+++ auction-server/src/auction/service/mod.rs
@@ -80,6 +80,7 @@ pub mod get_permission_keys_for_auction;
 pub mod handle_auction;
 pub mod handle_auctions;
 pub mod handle_bid;
+pub mod optimize_bids;
 pub mod simulator;
 pub mod submit_quote;
 pub mod update_bid_status;
@@ -229,7 +230,7 @@ impl ChainTrait for Svm {
 
 pub struct ServiceInner<T: ChainTrait> {
     opportunity_service: Arc<opportunity_service::Service<T::OpportunityServiceType>>,
-    pub config:          Config<T::ConfigType>,
+    config:              Config<T::ConfigType>,
     repo:                Arc<Repository<T>>,
     task_tracker:        TaskTracker,
     event_sender:        broadcast::Sender<UpdateEvent>,
