+++ auction-server/src/auction/service/simulator.rs
@@ -1,8 +1,4 @@
 use {
-    crate::{
-        auction::entities::Bid,
-        kernel::entities::Svm,
-    },
     futures::future::join_all,
     litesvm::{
         types::{
@@ -55,7 +51,7 @@ pub struct Simulator {
     account_cache: RwLock<HashMap<Pubkey, (Account, Instant)>>,
 }
 
-struct AccountsConfig {
+pub struct AccountsConfig {
     accounts:            HashMap<Pubkey, Account>,
     programs:            HashMap<Pubkey, Account>,
     upgradable_programs: HashMap<Pubkey, Account>,
@@ -257,7 +253,7 @@ impl Simulator {
     /// Uses the account cache to avoid fetching programs and lookup tables multiple times
     /// Returns an AccountsConfig struct that can be used to initialize the LiteSVM instance
     #[tracing::instrument(skip_all, fields(slot))]
-    async fn fetch_tx_accounts_via_rpc(
+    pub async fn fetch_tx_accounts_via_rpc(
         &self,
         transactions: &[VersionedTransaction],
     ) -> RpcResult<AccountsConfig> {
@@ -324,7 +320,10 @@ impl Simulator {
     }
 
     #[tracing::instrument(skip_all)]
-    fn setup_lite_svm(&self, accounts_config_with_context: &Response<AccountsConfig>) -> LiteSVM {
+    pub fn setup_lite_svm(
+        &self,
+        accounts_config_with_context: &Response<AccountsConfig>,
+    ) -> LiteSVM {
         let mut svm = LiteSVM::new()
             .with_sigverify(false)
             .with_blockhash_check(false)
@@ -398,39 +397,6 @@ impl Simulator {
         });
         let res = svm.simulate_transaction(transaction.clone());
         let res = Self::check_rent_exemption(&svm, res);
-        Ok(Response {
-            value:   res,
-            context: accounts_config_with_context.context,
-        })
-    }
-
-    /// Given a list of bids, tries to find the optimal set of bids that can be submitted to the chain
-    /// considering the current state of the chain and the pending transactions.
-    /// Right now, for simplicity, the method assume the bids are sorted, and tries to submit them in order
-    /// and only return the ones that are successfully submitted.
-    #[tracing::instrument(skip_all)]
-    pub async fn optimize_bids(&self, bids: &[Bid<Svm>]) -> RpcResult<Vec<Bid<Svm>>> {
-        let pending_txs = self.fetch_pending_and_remove_old_txs().await;
-        let txs_to_fetch = pending_txs
-            .iter()
-            .chain(bids.iter().map(|bid| &bid.chain_data.transaction))
-            .cloned()
-            .collect::<Vec<_>>();
-        let accounts_config_with_context = self.fetch_tx_accounts_via_rpc(&txs_to_fetch).await?;
-        let mut svm = self.setup_lite_svm(&accounts_config_with_context);
-
-        pending_txs.into_iter().for_each(|tx| {
-            let _ = svm.send_transaction(tx);
-        });
-        let mut res = vec![];
-        for bid in bids {
-            if svm
-                .send_transaction(bid.chain_data.transaction.clone())
-                .is_ok()
-            {
-                res.push(bid.clone());
-            }
-        }
         Ok(Response {
             value:   res,
             context: accounts_config_with_context.context,
